[Fix] Sample tab: Fix to crash that occurred when resizing sample while it was playing. [Fix] Instrument tab: Fix to crash that occurred when multiple windows had instrument tab of the same document open. [Fix] Tuning: Default tuning file should now load correctly (broken in rev. 323). [Ref] S3M load: Check parameters before reading. [Ref] Internal: Changes to get rid of some compiler warnings and disabled some warnings (see Stdafx.h).
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@327 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/ScaleEnvPointsDlg.h
@@ -11,7 +11,7 @@ class CScaleEnvPointsDlg : public CDialog
 	DECLARE_DYNAMIC(CScaleEnvPointsDlg)
 
 public:
-	CScaleEnvPointsDlg(CWnd* pParent, MODINSTRUMENT* pInst, BYTE env);   // standard constructor
+	CScaleEnvPointsDlg(CWnd* pParent, MODINSTRUMENT* pInst, UINT env);   // standard constructor
 	virtual ~CScaleEnvPointsDlg();
 
 // Dialog Data
@@ -24,7 +24,7 @@ protected:
 
 private:
 	MODINSTRUMENT* m_pInstrument;
-	BYTE m_Env; //To tell which envelope to process.
+	UINT m_Env; //To tell which envelope to process.
 	CEdit m_EditFactor;
 protected:
 	virtual void OnOK();
